+++ Frontend/src/components/ui/PlayerSlider/PlayerSlider.module.css
