+++ Frontend/src/components/ui/Slider/Slider.js
