+++ Frontend/src/components/ui/Slider/Slider.module.css
