+++ Frontend/src/components/ui/VolumeSlider/VolumeSlider.module.css
